+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 01. :gift: Added interface `dev.orne.beans.Identity`
 01. :gift: Added class `dev.orne.beans.TokenIdentity`
 01. :gift: Added exception `dev.orne.beans.UnrecognizedIdentityTokenException`
+01. :gift: Added class `dev.orne.beans.IdentityTokenFormatter`
 01. :gift: Added annotation `dev.orne.beans.IdentityTokenResolver`
 01. :gift: Added class `dev.orne.beans.IdentityResolver`
 01. :gift: Added interface `dev.orne.beans.IdentityBean`